fix(useThrottle): Support function values
If the caller passes a function as the value, this invokes the function form of
useState and setState which does not set the state to the value but instead invokes
the function and sets the state to the return value of the function. This causes
useThrottle to behave incorrectly when passing a value which is a function. Rather
than returning the function (as the type signature indicates), useThrottle returns
the return value of the function.

This changes fixes useThrottle to support function values as the value to throttle
to align with the type signature.