fix(useThrottle): add type annotation for values which are functions
If the value is a function, it is invoked and the return value of the function
becomes the throttled state. This happens because value is passed directly to
setState which has different behavior for function values.

This change updates the type annotation to correctly reflect the behavior
so users who pass function values get the correct return type (T) from
useThrottle.