fix(createBreakpoint): improve the return type
before this commit, when using the createBreakpoint() function with or without a custom breakpoint, it will always return a type `() => string`.

this commit adds overloaded and mapped types to the createBreakpoint function definition that reduces the cardinality of the type results according to the given parameters.

in case no custom breakpoints are given, the keys from the default object are resolved as the return type of the useBreakpoint function