Missing dependencies in hook in useLocalStorage().
The lack of these dependencies in the set useCallback hook on line 59 (especially the lack of the 'state' dependency) was causing incorrect behaviour.

E.g. when the set method is called with a function, it would use an out of date state and would therefore apply the set function to the old state, causing an unexpected result.