fix: Call the handler as soon as the event is listened to
useKeyPressEvent previously listened for changes in the state of the key through the useKeyPress hook instead of directly listening for the event itself. Switching the useKeyPress call to useKey and listening directly to the events would resolve this problem, but would slightly change the parameter type.

Closes: #201

BREAKING-CHANGE: Change the last parameter of useKeyPressEvent from `useKeyPress = useKeyPressDefault` to `useKey = useKeyDefault`